shared/python-common/pyproject.toml
[project]
name = "artisan-common"
version = "0.1.0"
description = "Shared library for Artisan platform microservices"
requires-python = ">=3.13"
dependencies = [
    "pydantic>=2.9,<3.0",
    "pydantic-settings>=2.6,<3.0",
    "fastapi>=0.115,<1.0",
    "opentelemetry-api>=1.28,<2.0",
    "opentelemetry-sdk>=1.28,<2.0",
    "opentelemetry-instrumentation-fastapi>=0.49b0",
    "structlog>=24.4,<25.0",
    "nats-py>=2.9,<3.0",
    "python-jose[cryptography]>=3.3,<4.0",
    "httpx>=0.27,<1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3,<9.0",
    "pytest-asyncio>=0.24,<1.0",
    "pytest-cov>=6.0,<7.0",
    "ruff>=0.7,<1.0",
    "mypy>=1.13,<2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/artisan_common"]

[tool.ruff]
target-version = "py313"
line-length = 120
src = ["src"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort (import ordering)
    "N",    # pep8-naming
    "UP",   # pyupgrade (modern Python syntax)
    "B",    # bugbear (common bugs)
    "S",    # bandit (security)
    "A",    # builtins shadowing
    "T20",  # print statements (should use logging)
    "RUF",  # ruff-specific rules
]
ignore = [
    "S101",  # assert usage (fine in tests)
    "S101",  # assert usage (fine in tests)
    "B008",  # function calls in defaults (required for FastAPI Depends)
]

[tool.ruff.lint.isort]
known-first-party = ["artisan_common"]

[tool.mypy]
python_version = "3.13"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --tb=short --strict-markers"
